added loader to confirmation window, moved files to subfolder
