+++ src/components/Profile/components/PersonalInfoList/components/AlumniUpdateForm/ConfirmationHeader/ConfirmationHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../vars';
+@import '../../../../../../../vars';
 
 %header_style {
   padding-left: 5px;
