+++ src/components/Profile/components/PersonalInfoList/components/AlumniUpdateForm/ConfirmationRow/ConfirmationRow.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../../vars';
+@import '../../../../../../../vars';
 
 %text {
   text-transform: capitalize;
