+++ src/components/Profile/components/PersonalInfoList/components/AlumniUpdateForm/Update.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../vars';
+@import '../../../../../../vars';
 
 .select_text {
   text-transform: capitalize;
