+++ src/routes.js
@@ -6,7 +6,6 @@ import CoCurricularTranscript from './views/CoCurricularTranscript';
 import EnrollmentCheckIn from './views/EnrollmentCheckIn';
 import Events from './views/Events';
 import EventsAttended from './views/EventsAttended';
-import AlumniUpdateForm from './views/AlumniUpdateForm';
 import Feedback from './views/Feedback';
 import Help from './views/Help';
 import Home from './views/Home';
